Upgrade Node packages
The big changes are to Babel, Webpack and Jest.

Jest was mostly some small changes to syntax. An issue with the Airtable library's use of fetch means that we need to use Jest with whatwg-fetch.

Webpack 5 and Babel 7.4 removed polyfills by default, so there are a couple of manual additions to fill in the gaps.