Remove secure property from setting parameters
Secure settings only work in requests made with the Zendesk framework's `client.request()` method which would make our code complex. Given this is a private app, the security risk is low.

Didn't catch this before because it only has effect when the app is deployed, not during testing.

The app ID has to be updated because to remove a secure property the app must be deleted and reinstalled.