Fix: point at correct code to resolve dependency
We are pointing at the v3.8.5 tag on dxw/zendesk_apps_tools which does
not have the updated dependencies.

Perhaps we should bump the version of dxw/zendesk_apps_tools  to v.3.8.6
with the updated Gemfile and then resolve the latest version here. I am
assuming the fork is for these updates as the actual project appears to
be in maintenance mode (!?!)

As I don't want to mess with multiple repos the lowest friction fix is
to point at the ref that has the fix in dxw/zendesk_apps_tools. This
will clear the security alerts we have on this repository.