Handle services that lack a service type
Some services like Open Hours don't have a service type. That means that we can't access the `data` attribute and an error gets thrown.

Converting the chaining operator to be optional means that instead of throwing an error we get an undefined response and can test the next item in the array.