Allow the form to handle multiple services
Previously time could only be logged against a single support service. Alex Farley requested that we support multiple support services and allow the user to select which one they want to log time against.

This commit adds support for that - adding a dropdown menu to the form, and pluralising references to 'service'.